Fixed test and add test about report
