+++ test/config/translator_report.yaml
@@ -1,0 +1,48 @@
+# Small config to help refactoring to make the model completely YAML based
+defaults:
+  experiment:
+    model_file: examples/output/<EXP>.mod
+    hyp_file: examples/output/<EXP>.hyp
+    out_file: examples/output/<EXP>.out
+    err_file: examples/output/<EXP>.err
+    run_for_epochs: 1
+    eval_metrics: bleu
+  train:
+    trainer: adam
+    learning_rate: 0.01
+    default_layer_dim: 32
+    dropout: 0.0
+    dev_metrics: bleu
+    training_corpus: !BilingualTrainingCorpus
+      train_src: examples/data/head.ja
+      train_trg: examples/data/head.en
+      dev_src: examples/data/head.ja
+      dev_trg: examples/data/head.en
+    corpus_parser: !BilingualCorpusParser
+      src_reader: !PlainTextReader {}
+      trg_reader: !PlainTextReader {}
+      max_src_len: 15
+      max_trg_len: 15
+    model: !DefaultTranslator
+      src_embedder: !SimpleWordEmbedder
+        emb_dim: 32
+      encoder: !LSTMEncoder
+        layers: 1
+        input_dim: 32
+      attender: !StandardAttender
+        state_dim: 32
+        hidden_dim: 32
+        input_dim: 32
+      trg_embedder: !SimpleWordEmbedder
+        emb_dim: 32
+      decoder: !MlpSoftmaxDecoder
+        layers: 1
+        mlp_hidden_dim: 32
+  decode:
+    src_file: examples/data/head.ja
+    report_path: examples/output/<EXP>.report
+    report_type: html, file
+  evaluate:
+    ref_file: examples/data/head.en
+
+debug:
